+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2025-08-26
+## [Unreleased] - 2025-09-09
+
+### Added
+
+-   Add zoom and reset view functionality to the dependency graph.
+
+### Changed
+
+-   Refactor CSS and JavaScript into separate files (`assets/styles.css` and `assets/app.js`).
+
+## [0.1.0] - 2025-08-26
 
 ### Added